Address PR feedback 2
